move files around for docker stuff
